CAMEL-23857: Move profile to DefaultConfigurationProperties so all runtimes (Spring Boot, Quarkus) can use it
